add some methods to accept the data from Nerd in form of WikidataElement (wikidata id, label, and properties), add a unit testing

+++ src/main/java/org/nerd/kid/arff/ArffFileGenerator.java
@@ -150,8 +150,8 @@ public class ArffFileGenerator {
     public ArffFileGenerator addSingle(WikidataElementInfos result) {
         StringBuilder line = new StringBuilder();
 
-        Integer[] features = result.getFeatureVector();
-        for (Integer feature : features) {
+        Double[] features = result.getFeatureVector();
+        for (Double feature : features) {
             line.append(feature.toString()).append(",");
         }
         line.append(result.getRealClass()).append("\n");
